Logic to prevent posting the same memory twice
We have a state database that keeps track of the memories that have been posted, so once we randomly select a memory check against the state database to see if it's already been posted. If so, pick a new one and check again.

Eventually this will be problematic, but by that point I'll probably implement a periodic process to reap old posted memories out of the state database so they get re-posted after some time. Definite optimizations that can come later here.