Moving content repos into main bot repo
I'm centralizing all of the content for the nostalgiabot into this repo. I used to keep everything in separate repos because I was storing the full dump of photos in full quality as git objects but that's just silly to keep doing and also prevented a lot of automated behavior I wanted to do eventually.

This copies all of the metadata from nostalgiabot-obama and nostalgiabot-biden into this repo. I'mm be deleting those repos once the code has been updated to use this new location.